src/main/java/com/mors741/spittr/web/SpittleController.java
package com.mors741.spittr.web;

import com.mors741.spittr.Spittle;
import com.mors741.spittr.data.SpittleRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;

import java.util.List;

@Controller
@RequestMapping(path = "/spittles")
public class SpittleController {

    private SpittleRepository spittleRepository;

    @Autowired
    public SpittleController(SpittleRepository spittleRepository) {
        this.spittleRepository = spittleRepository;
    }

    @RequestMapping(method = RequestMethod.GET)
    public List<Spittle> spittles(
            @RequestParam(name = "max", defaultValue = "9223372036854775807") long max,
            @RequestParam(name = "count", defaultValue = "20") int count
    ) {
        return spittleRepository.findSpittles(max, count);
    }
}
